+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -89,6 +89,12 @@ New checks
   Finds member initializations in the constructor body which can be placed into
   the initialization list instead.
 
+- New :doc:`misc-fmt-argument-mismatch
+  <clang-tidy/checks/misc-fmt-argument-mismatch>` check.
+
+  FIXME: add release notes.
+
+
 New check aliases
 ^^^^^^^^^^^^^^^^^